added lib.d.ts for types
I added this so that we will have consistencies with types and lessen type errors in passing of props
