+++ types.d.ts
@@ -1,0 +1,9 @@
+type MemberProp = {
+  id: string;
+  name: string;
+};
+
+type ChallengeProp = {
+  id: string;
+  name: string;
+};
